+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog].
 -   The `as_dict` method of Qobj is deprecated in favor of `to_dict`.
 
 ### Added
+-   Ability to check for equality of pulse `Schedule` and `Instruction`.
 -   Added tests for `gate_map` and reference images for testing `plot_gate_map`
 -   New `CountOpsLongest` analysis pass to retrieve the number of operations
     on the longest path of the DAGCircuit.
@@ -64,6 +65,7 @@ The format is based on [Keep a Changelog].
 
 ### Changed
 
+-   `Schedule.instructions` now returns with time-ordering.
 -   More informative errors are now raised if `qubit_lo_freq` and
     `meas_lo_freq` are not supplied to `assemble_schedules`.
 -   `pulse.samplers` module has now been moved to `pulse.pulse_lib.samplers`.